+++ extensions/vscode/src/language-server/index.ts
@@ -1,0 +1,2 @@
+export * from "./language-server";
+export * from "./selectors";
