+++ extensions/vscode/src/language-server/selectors.ts
@@ -1,0 +1,7 @@
+const ANALYSIS_OPTIONS_FILE = {
+  pattern: "**/analysis_options.yaml",
+  scheme: "file",
+};
+const DART_FILE = { language: "dart", scheme: "file" };
+
+export { ANALYSIS_OPTIONS_FILE, DART_FILE };
